Add detection of license from license file
For packagers that have little or no metadata, such as rebar when used
with git deps, we have to detect the license type from the license file.

The code is mostly ported from Pivotal's LicenseFinder under the MIT
license.

Additional license variants are added based on initial testing against a
proprietary erlang project.

1. https://github.com/pivotal/LicenseFinder