Add perl/cpan support
A Note on the design:

Perl packaging lacks a lot of niceties as far as the task of license
collection goes. For Chef Software products, we use `cpanm` to install
things, which is similar to `gem install foo` in that the install just
puts everything in your global library location without any indication
of what was installed.

On top of that, perl dependencies are expressed in terms of modules, but
the unit of packaging/distribution is a tarball called a distribution.
There is a convention for how the names of the two things map to each
other, but not every package follows it.

Furthermore, dists mostly have decent metadata, but the install process
doesn't leave a copy of the full dist you installed, so that information
gets lost.

Finally, most of the available tooling to resolve the dependencies of a
given lib/app are not great for our purposes; some of the failings
include: requiring you to run the program, relying on a lot of
dependencies, relying on non-pure dependencies, and making requests to
web services for dependency information.

Therefore, this implementation:
1. queries http://deps.cpantesters.org to get a list of dependencies for
   the given software
2. parses the response to get a list of cpan dists
3. caches a copy of the dist locally
4. unpacks the tarball
5. inspects the content for metadata and license files

Of particular concern is that the deps.cpantesters site can be quite
slow to respond when requesting the dependency list for applications
with a large number of deps and we don't know how reliable it will be.
That said, it was decided that it's a mildly preferable dev experience
to click "rebuild" rather than attempt to pre-cache the dependency info
and possibly need manual intervention when deps become stale.

Unfortunately, there are not good alternatives that don't involve
modifying the way the application is installed (e.g., to use `carton`,
perl's version of bundler) or switching to a non-perl alternative
software.