Add Habitat support
This makes it so you can scan habitat packages and their dependencies.
It should work, however, it's only going to have `license_files` if a
package contains a file called LICENSE (see habitat-sh/habitat#1738),
which most currently don't.

This is only a first step, and does not take into account "internal"
dependencies, like if a package is built for a Node or Ruby project.
This kind of stuff can be built into the build process for Habitat
packages.

Signed-off-by: Nathan L Smith <smith@chef.io>