Fix bad link for Artistic License
The perl foundation rearranged their website, and the link
(http://www.perlfoundation.org/attachment/legal/artistic-2_0.txt) for
the Artistic License went away.

There's a html formatted page here:
https://www.perlfoundation.org/artistic-license-20.html. However, on
examination of the actual sources, most of them pointed elsewhere.

There are two patterns in the exceptions file for license files, and
I'm not sure if I hit the right one. One pattern is to cite the source
file that names the license. Another is to list an external file that
contains the text of the actual license. I've chosen to do the first.

Signed-off-by: Mark Anderson <mark@chef.io>