Make hab dep manager match hab's channel fallback behavior
Habitat will install a package from a specified channel instead of
"stable" if so configured, but if it cannot find a package in that
channel it will fallback to "stable", or just install the package from
unstable if a fully qualified pkg ident is provided. This updates the
habitat dependency manager logic in license scout to match.

Signed-off-by: Daniel DeLeo <dan@chef.io>