Add flag that allows you to easily search all subdirectories
There may be situations where it is undesirable or unreasonable to
individually specify all the directories you want to scan. This change
adds the `--include-all-subdirectories` flag which will tell license
scout to also look for dependency manager files in nested
subdirectories of the configured directory.

In practice, you could run the following command in the root of a
project repository and feel pretty confident that you have indentified
all of the dependencies of all of the software in that project.

```
license_scout --include-sub-directories
```

Signed-off-by: Tom Duffield <tom@chef.io>